+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -21,6 +21,7 @@ from rana_qgis_plugin.utils import (
 )
 from rana_qgis_plugin.utils_api import (
     get_tenant_file_descriptor,
+    get_tenant_project_file,
     get_tenant_project_files,
     get_tenant_projects,
     get_threedi_schematisation,
@@ -544,10 +545,18 @@ class RanaBrowser(uicls, basecls):
         assert isinstance(sender, QThread)
         sender.file_overwrite = file_overwrite
 
+    def refresh_file_data(self):
+        self.communication.bar_info("Refreshing local file references")
+        self.selected_file = get_tenant_project_file(
+            self.project["id"], {"path": self.selected_file["id"]}
+        )
+        self.communication.bar_info("Local file references refreshed")
+
     def on_file_upload_finished(self):
         self.rana_widget.setEnabled(True)
         self.communication.clear_message_bar()
         self.communication.bar_info(f"File uploaded to Rana successfully!")
+        self.refresh_file_data()
         sender = self.sender()
         assert isinstance(sender, QThread)
         sender.quit()