+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -562,12 +562,17 @@ class RanaBrowser(uicls, basecls):
             f"{self.project['name']}/{self.selected_file['id']}/last_modified"
         )
         QSettings().setValue(last_modified_key, self.selected_file["last_modified"])
+        self.show_selected_file_details()
 
     def on_file_upload_finished(self):
         self.rana_widget.setEnabled(True)
         self.communication.clear_message_bar()
         self.communication.bar_info(f"File uploaded to Rana successfully!")
-        self.refresh_file_data()
+        # show updated file data if file is selected, new list of files if not
+        if self.selected_file:
+            self.refresh_file_data()
+        else:
+            self.fetch_and_populate_files()
         sender = self.sender()
         assert isinstance(sender, QThread)
         sender.wait()
@@ -620,7 +625,11 @@ class RanaBrowser(uicls, basecls):
 
     def on_vector_style_finished(self, msg: str):
         self.rana_widget.setEnabled(True)
-        self.refresh_file_data()
+        # show updated file data if file is selected, new list of files if not
+        if self.selected_file:
+            self.refresh_file_data()
+        else:
+            self.fetch_and_populate_files()
         self.communication.clear_message_bar()
         self.communication.show_info(msg)