+++ rana_qgis_plugin/workers.py
@@ -8,7 +8,7 @@ import requests
 from PyQt5.QtCore import QSettings, QThread, pyqtSignal, pyqtSlot
 from qgis.core import Qgis, QgsMessageLog, QgsProject
 
-from .libs.bridgestyle.mapboxgl.fromgeostyler import convertGroup
+from bridgestyle.mapboxgl.fromgeostyler import convertGroup
 from .utils import get_local_file_path, image_to_bytes
 from .utils_api import (
     finish_file_upload,