+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -79,8 +79,12 @@ class RanaBrowser(uicls, basecls):
         self.projects_tv.header().sortIndicatorChanged.connect(self.sort_projects)
         self.projects_tv.clicked.connect(self.select_project)
         self.projects_search.textChanged.connect(self.filter_projects)
-        self.refresh_btn.setIcon(refresh_icon)
-        self.refresh_btn.clicked.connect(self.refresh_projects)
+        self.overview_refresh_btn.setIcon(refresh_icon)
+        self.overview_refresh_btn.clicked.connect(self.refresh)
+        self.project_refresh_btn.setIcon(refresh_icon)
+        self.project_refresh_btn.clicked.connect(self.refresh)
+        self.file_refresh_btn.setIcon(refresh_icon)
+        self.file_refresh_btn.clicked.connect(self.refresh)
         self.fetch_projects()
         self.populate_projects()
         self.projects_tv.header().setSortIndicator(1, Qt.AscendingOrder)
@@ -325,16 +329,18 @@ class RanaBrowser(uicls, basecls):
         self.selected_file = file_item.data(Qt.UserRole)
         self._update_file_UI()
 
-    def _update_file_UI(self):
+    def _update_file_UI(self, append_path: bool = True):
         file_path = self.selected_file["id"]
         if self.selected_file["type"] == "directory":
             directory_name = os.path.basename(file_path.rstrip("/"))
-            self.paths.append(directory_name)
+            if append_path:
+                self.paths.append(directory_name)
             self.fetch_and_populate_files(file_path)
             self.rana_widget.setCurrentIndex(1)
         else:
             file_name = os.path.basename(file_path.rstrip("/"))
-            self.paths.append(file_name)
+            if append_path:
+                self.paths.append(file_name)
             self.show_selected_file_details()
             self.rana_widget.setCurrentIndex(2)
 
@@ -610,6 +616,7 @@ class RanaBrowser(uicls, basecls):
         sender.file_overwrite = file_overwrite
 
     def refresh_file_data(self):
+        assert self.selected_file
         self.selected_file = get_tenant_project_file(
             self.project["id"], {"path": self.selected_file["id"]}
         )
@@ -617,19 +624,30 @@ class RanaBrowser(uicls, basecls):
             f"{self.project['name']}/{self.selected_file['id']}/last_modified"
         )
         QSettings().setValue(last_modified_key, self.selected_file["last_modified"])
+        self._update_file_UI(append_path=False)
 
-    def on_file_upload_finished(self, refresh: bool = True):
+    def refresh(self):
+        current_index = self.rana_widget.currentIndex()
+        if current_index == 0:
+            self.refresh_projects()
+        elif current_index == 1:
+            self.fetch_and_populate_files()
+        elif current_index == 2:
+            self.refresh_file_data()
+        else:
+            raise Exception("cannot refresh; rana_widget index must be 1, 2, or 3")
+
+    def on_file_upload_finished(self):
         self.rana_widget.setEnabled(True)
         self.communication.clear_message_bar()
         self.communication.bar_info(f"File uploaded to Rana successfully!")
-        if refresh:
-            self.refresh_file_data()
+        self.refresh()
         sender = self.sender()
         assert isinstance(sender, QThread)
         sender.wait()
 
     def on_new_file_upload_finished(self, online_path: str):
-        self.on_file_upload_finished(False)
+        self.on_file_upload_finished()
         if self.communication.ask(
             self, "Load", "Would you like to load the uploaded file from Rana?"
         ):
@@ -691,7 +709,7 @@ class RanaBrowser(uicls, basecls):
 
     def on_vector_style_finished(self, msg: str):
         self.rana_widget.setEnabled(True)
-        self.refresh_file_data()
+        self.refresh()
         self.communication.clear_message_bar()
         self.communication.show_info(msg)