+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 
 - Added "Add File to Rana" button (nens/rana#1951)
 - Upload new file: set last used directory in settings (nens/rana#2058)
+- Reload local file url and other metadata after uploading to Rana.
 
 
 1.1.8 (2025-06-16)