+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 - Add option to pass Rana URL to QGIS. (nens/rana-qgis-plugin#61)
 - Add option to load WMS layers of 3Di scenarios (nens/rana-qgis-plugin#51)
 - Add option to download scenario zip (nens/rana-qgis-plugin#51)
+- Remove GeoCat/bridgestyle from the plugin since it is now on PyPI. (nens/rana-qgis-plugin#71)
 
 
 1.1.8 (2025-06-16)