+++ HISTORY.rst
@@ -14,6 +14,7 @@ History
 - Add option to download scenario zip (nens/rana-qgis-plugin#51)
 - Reload file ui after uploading file. (nens/rana#41)
 - Add refresh button to project and file overview to allow manual reloading.
+- Remove GeoCat/bridgestyle from the plugin since it is now on PyPI. (nens/rana-qgis-plugin#71)
 
 
 1.1.8 (2025-06-16)